Created RandomNumberCard.js to break code out
